requirements.txt
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
cffi==2.0.0
cryptography==46.0.3
deprecation==2.1.0
distro==1.9.0
et_xmlfile==2.0.0
greenlet==3.2.4
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
jiter==0.11.1
multidict==6.7.0
numpy==2.3.4
openai==2.6.1
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
playwright==1.55.0
postgrest==2.23.0
propcache==0.4.1
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
pyee==13.0.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
realtime==2.23.0
six==1.17.0
sniffio==1.3.1
storage3==2.23.0
StrEnum==0.4.15
supabase==2.23.0
supabase-auth==2.23.0
supabase-functions==2.23.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
websockets==15.0.1
yarl==1.22.0
